+++ package.json
@@ -12,8 +12,10 @@
     "binary-release": "node ./scripts/binary.js release",
     "binary-upload": "node ./scripts/binary.js upload",
     "binary-zip": "node ./scripts/binary.js zip",
-    "build": "lerna run build --stream --no-bail --ignore create-cypress-tests --ignore cypress  --ignore \"'@packages/{runner}'\" && lerna run build --stream --scope create-cypress-tests --scope cypress",
-    "build-prod": "lerna run build-prod-ui --stream && lerna run build-prod --stream --ignore create-cypress-tests --ignore cypress && lerna run build-prod --stream --scope create-cypress-tests --scope cypress",
+    "build": "lerna run build --stream --no-bail --ignore create-cypress-tests --ignore \"'@packages/{runner}'\"",
+    "postbuild": "lerna run build --stream --scope create-cypress-tests && node ./cli/scripts/post-main-build.js",
+    "build-prod": "lerna run build-prod-ui --stream && lerna run build-prod --stream --ignore create-cypress-tests",
+    "postbuild-prod": "lerna run build-prod --stream --scope create-cypress-tests && node ./cli/scripts/post-main-build.js",
     "check-node-version": "node scripts/check-node-version.js",
     "check-terminal": "node scripts/check-terminal.js",
     "clean": "lerna run clean --parallel --no-bail || echo 'ok, errors while cleaning'",
